Added Cognito post-confirmation trigger to create an entry for new user in favorites database
